refactor to allow go lib openai with weaviate
