+++ backend/go-api/openaiwrapper/chatmodel/chat_completion_model.go
