+++ backend/go-api/openaiwrapper/chatmodel/chat_completion_model_unit_test.go
