+++ backend/go-api/openaiwrapper/chatthread/chat_thread_unit_test.go
@@ -1,7 +1,7 @@
 package chatthread
 
 import (
-	"openai-integrations/openai/chatmodel"
+	"openaiwrapper-integrations/openaiwrapper/chatmodel"
 	"testing"
 )
