+++ backend/go-api/openaiwrapper/handlers/handlers.go
@@ -1,4 +1,4 @@
-package openai
+package openaiwrapper
 
 import (
 	"encoding/json"
