+++ backend/go-api/openaiwrapper/handlers/handlers_unit_test.go
@@ -1,4 +1,4 @@
-package openai
+package openaiwrapper
 
 import (
 	"bytes"
