+++ backend/go-api/openaiwrapper/sales_utils.go
@@ -1,4 +1,4 @@
-package openai
+package openaiwrapper
 
 func scheduleSalesCall() {
 	panic("unimplemented")
